+++ controllers/payment.controllers.js
@@ -4,6 +4,7 @@ const { v4: uuidv4 } = require('uuid')
 const msgClass = require('../errors/error')
 const paymentService = require('../services/payment.services')
 const paymentModel = require('../models/payment.models')
+const { isEmpty, doSomeAsyncMagik } = require('../utils/utils')
 
 
 const createTransaction = async (req, res) => {
@@ -139,10 +140,218 @@ const verifyTransaction = async (req, res) => {
  
  }
 
+
+ const createNewCustomer = async (req,res) => {
+
+    const{ email, first_name, last_name, phone} = req.body
+
+    const customerSchema = Joi.object({
+        email: Joi.string().email().required(),
+        first_name: Joi.string().required(),
+        last_name: Joi.string().required(),
+        phone: Joi.string()
+    })
+    try {
+    const responseFromJoiValidation = customerSchema.validate(req.body)
+        if(responseFromJoiValidation.error) {
+            throw new Error("Bad request")
+        }
+        const createCustomerResponse = await paymentService.createCustomer(req.body)
+        console.log("Got back from paysatck: ", JSON.stringify(createCustomerResponse.data))
+        if (createCustomerResponse.data.status ==false) {
+            throw new Error("Customer hasn't been created")
+        }
+
+        res.status(200).send({
+            status: true,
+            message: "Customer created",
+            data: createCustomerResponse.data.data
+        })  
+    }
+    catch(error){
+        res.status(400).send({
+            status:false,
+            message: error.message || msgClass.GeneralError
+        })
+    }
+
+}
+
+
+const listCustomers = async (req,res) => {
+    const perPage = parseInt(req.query.perPage) || 50
+    const page = parseInt(req.query.page) || 1
+
+    try{
+        const getCustomersListResponse = await paymentService.getCustomersList(perPage, page)
+        console.log('i got here')
+        if (getCustomersListResponse.data.status != true){
+            throw new Error("Customer list could not be fetched")
+        }
+
+        res.status(200).send({
+            status: true,
+            message: "Customers recieved",
+            data: getCustomersListResponse.data.data
+        })
+    }
+    catch(e) {
+        // console.log(`error: ${e.message}`)
+         res.status(400).send({
+             status: false,
+             message: e.message || msgClass.GeneralError
+ 
+      })
+     }
+}
+
+const getCustomer = async (req, res) => {
+
+     const { email } = req.params
+    
+     try {
+ 
+         const fetchCustomerResponse = await paymentService.fetchCustomer(email)
+         console.log(fetchCustomerResponse.data)
+         if (fetchCustomerResponse.data.status != true) {
+             throw new Error("We could not fetch Customer")
+         }
+         
+         res.status(200).send({
+             status: true,
+             message: "Customer retrieved",
+             data: fetchCustomerResponse.data.data
+         })
+     } 
+     catch(error) {
+        // console.log(`error: ${e.message}`)
+         res.status(400).send({
+             status: false,
+             message:   error.message || msgClass.GeneralError
+ 
+      })
+     }
+     
+     
+}     
+
+
+
+const customerUpdate = async (req,res) => {
+
+    // const{id_} = req.params
+
+    const{id, first_name, last_name} = req.body
+
+    const customerSchema = Joi.object({
+        first_name: Joi.string().required(),
+        last_name: Joi.string().required(),
+        
+    })
+    try {
+    const responseFromJoiValidation = customerSchema.validate(req.body)
+        if(responseFromJoiValidation.error) {
+            throw new Error("Bad request")
+        }
+        const updateCustomerResponse = await paymentService.updateCustomer(req.body)
+        if (updateCustomerResponse.data.status ==false) {
+            throw new Error("Customer could not update")
+        }
+
+        res.status(200).send({
+            status: true,
+            message: "Customer updated",
+            data: updateCustomerResponse.data.data
+        })  
+    }
+    catch(error){
+        res.status(400).send({
+            status:false,
+            message: error.message || msgClass.GeneralError
+        })
+    }
+
+}
+
+
+
+
+
+const whiteOrBlackListCustomer = async (req,res) => {
+
+    const { email , risk_action} = req.body
+
+    const whiteListschma = Joi.object({
+        email: Joi.string().email().required(),
+
+        risk_action: Joi.string().valid("default", "allow", "deny").required()
+    })
+    // paymentService.blackOrWhiteListing(req.body)
+    // .then(result => {
+    //     console.log(`result happened: `, result)
+    //     res.status(200).send({
+    //                     status: true,
+    //                     message: "succcess",
+    //                     data: result
+    //         }) 
+    // })
+    // .catch(err => {
+    //     console.log(`error happened: `, err)
+    //     res.status(422).send({
+    //                     status:false,
+    //                     message: err
+    //                 })
+    // })
+
+try{
+    const responseFromJoiValidation = whiteListschma.validate(req.body)
+    console.log("i was able to validate")
+        if (responseFromJoiValidation.error){
+            throw new Error("Invalid input")
+        }
+    
+        console.log("i am gpoing to call the service")
+        const  [err, responseWhiteBlacklistCustomer] = await doSomeAsyncMagik(paymentService.blackOrWhiteListing(req.body))
+        if (err){
+          //  console.log("error: ", err)
+            throw new Error("Something happened")
+
+        }
+        if (!responseWhiteBlacklistCustomer){
+            //  console.log("error: ", err)
+              throw new Error("it happened not well")
+  
+          }
+        console.log("i got a response", responseWhiteBlacklistCustomer)
+
+        res.status(200).send({
+            status: true,
+            message: responseWhiteBlacklistCustomer.data.message,
+            data: responseWhiteBlacklistCustomer
+        })  
+    
+    }
+    catch(error){
+        res.status(422).send({
+            status:false,
+            message: error
+        })
+    }
+}
+    
+
+
+
+
  
 
 
 module.exports = {
     createTransaction,
-    verifyTransaction
+    verifyTransaction,
+    createNewCustomer,
+    listCustomers,
+    getCustomer,
+    customerUpdate,
+    whiteOrBlackListCustomer
 }
